+++ dd-java-agent/instrumentation/vertx-pg-client-4/build.gradle
@@ -5,7 +5,7 @@ muzzle {
   pass {
     group = 'io.vertx'
     module = 'vertx-pg-client'
-    versions = '[4.2.0,5)'
+    versions = '[4.0.0,5)'
     assertInverse = true
   }
 }
@@ -20,9 +20,9 @@ latestDepTest {
 }
 
 dependencies {
-  compileOnly group: 'io.vertx', name: 'vertx-pg-client', version: '4.2.0'
+  compileOnly group: 'io.vertx', name: 'vertx-pg-client', version: '4.0.0'
 
-  testImplementation group: 'io.vertx', name: 'vertx-pg-client', version: '4.2.0'
+  testImplementation group: 'io.vertx', name: 'vertx-pg-client', version: '4.0.0'
   testImplementation project(':dd-java-agent:instrumentation:vertx-sql-common')
 
   // This is needed for the test container to start
