+++ dd-java-agent/instrumentation/vertx-pg-client-4/src/main/java/datadog/trace/instrumentation/vertx_pg_client_4/PgConnectionImplInstrumentation.java
@@ -1,4 +1,4 @@
-package datadog.trace.instrumentation.vertx_pg_client_4_2_0;
+package datadog.trace.instrumentation.vertx_pg_client_4;
 
 import static datadog.trace.agent.tooling.bytebuddy.matcher.NameMatchers.named;
 import static net.bytebuddy.matcher.ElementMatchers.isConstructor;
