+++ dd-java-agent/instrumentation/vertx-pg-client-4/src/main/java/datadog/trace/instrumentation/vertx_pg_client_4/PgPoolImplInstrumentation.java
@@ -1,38 +1,41 @@
-package datadog.trace.instrumentation.vertx_pg_client_4_2_0;
+package datadog.trace.instrumentation.vertx_pg_client_4;
 
 import static datadog.trace.agent.tooling.bytebuddy.matcher.NameMatchers.named;
+import static java.util.Collections.singletonMap;
 import static net.bytebuddy.matcher.ElementMatchers.*;
 
 import com.google.auto.service.AutoService;
 import datadog.trace.agent.tooling.Instrumenter;
 import datadog.trace.agent.tooling.InstrumenterModule;
-import java.util.HashMap;
+import datadog.trace.bootstrap.instrumentation.jdbc.DBInfo;
 import java.util.Map;
 
 @AutoService(InstrumenterModule.class)
-public class PreparedStatementImplInstrumentation extends InstrumenterModule.Tracing
+public class PgPoolImplInstrumentation extends InstrumenterModule.Tracing
     implements Instrumenter.ForSingleType, Instrumenter.HasMethodAdvice {
-  public PreparedStatementImplInstrumentation() {
+  public PgPoolImplInstrumentation() {
     super("vertx", "vertx-sql-client");
   }
 
   @Override
   public Map<String, String> contextStore() {
-    Map<String, String> contextStores = new HashMap<>();
-    contextStores.put("io.vertx.sqlclient.PreparedStatement", "datadog.trace.api.Pair");
-    contextStores.put("io.vertx.sqlclient.Query", "datadog.trace.api.Pair");
-    return contextStores;
+    return singletonMap("io.vertx.sqlclient.SqlClient", DBInfo.class.getName());
   }
 
   @Override
   public String instrumentedType() {
-    return "io.vertx.sqlclient.impl.PreparedStatementImpl";
+    return "io.vertx.pgclient.impl.PgPoolImpl";
   }
 
   @Override
   public void methodAdvice(MethodTransformer transformer) {
     transformer.applyAdvice(
-        isMethod().and(isPublic()).and(named("query")).and(takesNoArguments()),
-        packageName + ".PreparedStatementQueryAdvice");
+        isStatic()
+            .and(isPublic())
+            .and(isMethod())
+            .and(named("create"))
+            .and(takesArguments(4))
+            .and(takesArgument(2, named("io.vertx.pgclient.PgConnectOptions"))),
+        packageName + ".PgPoolImplAdvice");
   }
 }
