+++ dd-java-agent/instrumentation/vertx-pg-client-4/src/main/java/datadog/trace/instrumentation/vertx_pg_client_4/PreparedStatementImplInstrumentation.java
@@ -1,38 +1,38 @@
-package datadog.trace.instrumentation.vertx_pg_client_4_2_0;
+package datadog.trace.instrumentation.vertx_pg_client_4;
 
 import static datadog.trace.agent.tooling.bytebuddy.matcher.NameMatchers.named;
-import static java.util.Collections.singletonMap;
 import static net.bytebuddy.matcher.ElementMatchers.*;
 
 import com.google.auto.service.AutoService;
 import datadog.trace.agent.tooling.Instrumenter;
 import datadog.trace.agent.tooling.InstrumenterModule;
-import datadog.trace.bootstrap.instrumentation.jdbc.DBInfo;
+import java.util.HashMap;
 import java.util.Map;
 
 @AutoService(InstrumenterModule.class)
-public class PgConnectionFactoryInstrumentation extends InstrumenterModule.Tracing
+public class PreparedStatementImplInstrumentation extends InstrumenterModule.Tracing
     implements Instrumenter.ForSingleType, Instrumenter.HasMethodAdvice {
-  public PgConnectionFactoryInstrumentation() {
+  public PreparedStatementImplInstrumentation() {
     super("vertx", "vertx-sql-client");
   }
 
   @Override
   public Map<String, String> contextStore() {
-    return singletonMap("io.vertx.pgclient.impl.PgConnectionFactory", DBInfo.class.getName());
+    Map<String, String> contextStores = new HashMap<>();
+    contextStores.put("io.vertx.sqlclient.PreparedStatement", "datadog.trace.api.Pair");
+    contextStores.put("io.vertx.sqlclient.Query", "datadog.trace.api.Pair");
+    return contextStores;
   }
 
   @Override
   public String instrumentedType() {
-    return "io.vertx.pgclient.impl.PgConnectionFactory";
+    return "io.vertx.sqlclient.impl.PreparedStatementImpl";
   }
 
   @Override
   public void methodAdvice(MethodTransformer transformer) {
     transformer.applyAdvice(
-        isConstructor()
-            .and(takesArguments(2))
-            .and(takesArgument(1, named("io.vertx.pgclient.PgConnectOptions"))),
-        packageName + ".PgConnectionFactoryConstructorAdvice");
+        isMethod().and(isPublic()).and(named("query")).and(takesNoArguments()),
+        packageName + ".PreparedStatementQueryAdvice");
   }
 }
