+++ settings.gradle
@@ -502,7 +502,7 @@ include ':dd-java-agent:instrumentation:velocity'
 include ':dd-java-agent:instrumentation:vertx-mysql-client-3.9'
 include ':dd-java-agent:instrumentation:vertx-mysql-client-4.0'
 include ':dd-java-agent:instrumentation:vertx-mysql-client-4.4.2'
-include ':dd-java-agent:instrumentation:vertx-pg-client-4.2.0'
+include ':dd-java-agent:instrumentation:vertx-pg-client-4'
 include ':dd-java-agent:instrumentation:vertx-redis-client-3.9'
 include ':dd-java-agent:instrumentation:vertx-redis-client-3.9:stubs'
 include ':dd-java-agent:instrumentation:vertx-rx-3.5'
